[IMP] open_academy: Added multiple views in Session T#59063
Added 2 decorations for changing colors depending of the duration of the session.

If the duration of the session is less than 120, it is colored blue.
If the duration is greater than 360, it is colored red.

Added calendar view, a graph bar to see the number of attendees, and a kanban.

There were added 2 filters in course view:
A filter selected by default to show only the courses where you are the responsible
The second is to group by responsible